Implement router engine using a different approach
