+++ lib/layer.js
@@ -12,7 +12,6 @@
  * @private
  */
 
-var pathRegexp = require('path-to-regexp')
 var debug = require('debug')('router:layer')
 
 /**
@@ -28,23 +27,13 @@ var hasOwnProperty = Object.prototype.hasOwnProperty
 
 module.exports = Layer
 
-function Layer(path, options, fn) {
-  if (!(this instanceof Layer)) {
-    return new Layer(path, options, fn)
-  }
-
+function Layer(path, match, fn) {
   debug('new %s', path)
-  var opts = options || {}
 
+  this.path = path
   this.handle = fn
   this.name = fn.name || '<anonymous>'
-  this.params = undefined
-  this.path = undefined
-  this.regexp = pathRegexp(path, this.keys = [], opts)
-
-  if (path === '/' && opts.end === false) {
-    this.regexp.fast_slash = true
-  }
+  this.match = match
 }
 
 /**
@@ -93,83 +82,5 @@ Layer.prototype.handle_request = function handle(req, res, next) {
     fn(req, res, next)
   } catch (err) {
     next(err)
-  }
-}
-
-/**
- * Check if this route matches `path`, if so
- * populate `.params`.
- *
- * @param {String} path
- * @return {Boolean}
- * @api private
- */
-
-Layer.prototype.match = function match(path) {
-  if (path == null) {
-    // no path, nothing matches
-    this.params = undefined
-    this.path = undefined
-    return false
-  }
-
-  if (this.regexp.fast_slash) {
-    // fast path non-ending match for / (everything matches)
-    this.params = {}
-    this.path = ''
-    return true
-  }
-
-  var m = this.regexp.exec(path)
-
-  if (!m) {
-    this.params = undefined
-    this.path = undefined
-    return false
-  }
-
-  // store values
-  this.params = {}
-  this.path = m[0]
-
-  // iterate matches
-  var keys = this.keys
-  var params = this.params
-
-  for (var i = 1; i < m.length; i++) {
-    var key = keys[i - 1]
-    var prop = key.name
-    var val = decode_param(m[i])
-
-    if (val !== undefined || !(hasOwnProperty.call(params, prop))) {
-      params[prop] = val
-    }
-  }
-
-  return true
-}
-
-/**
- * Decode param value.
- *
- * @param {string} val
- * @return {string}
- * @private
- */
-
-function decode_param(val){
-  if (typeof val !== 'string' || val.length === 0) {
-    return val
-  }
-
-  try {
-    return decodeURIComponent(val)
-  } catch (err) {
-    if (err instanceof URIError) {
-      err.message = 'Failed to decode param \'' + val + '\''
-      err.status = 400
-    }
-
-    throw err
   }
 }
