+++ lib/route.js
@@ -14,8 +14,8 @@
 
 var debug = require('debug')('router:route')
 var flatten = require('array-flatten')
-var Layer = require('./layer')
 var methods = require('methods')
+var Layer = require('./layer')
 
 /**
  * Module variables.
@@ -31,7 +31,7 @@ var slice = Array.prototype.slice
 module.exports = Route
 
 /**
- * Initialize `Route` with the given `path`,
+ * Initialize `Route` with the given `path`.
  *
  * @param {String} path
  * @api private
@@ -182,7 +182,7 @@ Route.prototype.all = function all(handler) {
       throw new TypeError('argument handler must be a function')
     }
 
-    var layer = Layer('/', {}, fn)
+    var layer = new Layer('/', null, fn)
     layer.method = undefined
 
     this.methods._all = true
@@ -209,7 +209,7 @@ methods.forEach(function (method) {
 
       debug('%s %s', method, this.path)
 
-      var layer = Layer('/', {}, fn)
+      var layer = new Layer('/', null, fn)
       layer.method = method
 
       this.methods[method] = true
